Add redux proof-of-concept
