+++ flask_server.py
@@ -4,7 +4,7 @@ import sys
 import os
 from os.path import join as pjoin
 import tarfile
-from json_util import to_json
+
 from flask import (
     Flask, request, abort, render_template,
     session, Response, jsonify, g, send_from_directory)
