+++ public/scripts/main.jsx
@@ -1,21 +1,45 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
-import Modal from 'react-modal';
-import FileInput from 'react-file-input';
-import ReactTabs from 'react-tabs';
-import CheckboxGroup from 'react-checkbox-group';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import ReactCSSTransitionGroup from 'react-addons-css-transition-group'
+import FileInput from 'react-file-input'
+import ReactTabs from 'react-tabs'
+import CheckboxGroup from 'react-checkbox-group'
 
 var Tab = ReactTabs.Tab;
 var Tabs = ReactTabs.Tabs;
 var TabList = ReactTabs.TabList;
 var TabPanel = ReactTabs.TabPanel;
 
-import _ from 'underscore';
-import filter from 'filter-values';
+import _ from 'underscore'
+import filter from 'filter-values'
 
-import 'bootstrap-css';
-import 'bootstrap';
+import 'bootstrap-css'
+import 'bootstrap'
+
+
+import { Provider } from 'react-redux'
+import configureStore from './configureStore'
+
+import * as Action from './actions';
+
+
+const store = configureStore()
+
+
+import ProjectList from './ProjectList'
+import { FormInputRow, FormSelectInput, FormTitleRow } from './Form'
+
+
+function json_post(url, body) {
+  return fetch('/newProject', {
+    method: 'POST',
+    headers: {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    },
+    body: body,
+  });
+}
 
 
 var MainContent = React.createClass({
@@ -63,25 +87,7 @@ var MainContent = React.createClass({
         this.loadState();
     },
     loadState: function() {
-        $.ajax({
-            url: "/get_state",
-            dataType: "json",
-            cache: false,
-            success: function(data) {
-                this.setState(
-                    {
-                        projectsList: data.projectsList,
-                        datasetsList: data.datasetsList,
-                        modelsList: data.modelsList,
-                        featuresetList: data.featuresetList,
-                        predictionsList: data.predictionsList
-                    });
-            }.bind(this),
-            error: function(xhr, status, err) {
-                console.error("/get_state", status, err.toString(),
-                              xhr.repsonseText);
-            }.bind(this)
-        });
+      store.dispatch(Action.fetchProjects());
     },
     updateProjectList: function() {
         $.ajax({
@@ -380,229 +386,6 @@ var NewProjectForm = React.createClass({
     }
 });
 
-var ProjectList = React.createClass({
-    render: function() {
-        var projectNodes = this.props.projectsList.map(function(project) {
-            return (
-                <ProjectListRow
-                    project={project}
-                    key={project.id}
-                    clickEditProject={this.props.clickEditProject}
-                    deleteProject={this.props.deleteProject}
-                    projectDetails={this.props.projectDetails}
-                    handleInputChange={this.props.handleInputChange}
-                    updateProjectInfo={this.props.updateProjectInfo}
-                />
-            );
-        }.bind(this));
-        return (
-            <div className="projectListDiv" style={{marginTop: 40}}>
-                <h3>Existing Projects</h3>
-                <div style={{width: 320, float: 'left'}}>
-                    <b>Name</b>
-                </div>
-                <div style={{marginLeft: 20, width: 320, float: 'left'}}>
-                    <b>Date Created</b>
-                </div>
-                <div style={{marginLeft: 710}}>
-                    <b>Edit/Delete Project</b>
-                </div>
-                <div>
-                    <ReactCSSTransitionGroup
-                        transitionName="projectsListItems"
-                        transitionEnterTimeout={200}
-                        transitionLeaveTimeout={200}>
-                        {projectNodes}
-                    </ReactCSSTransitionGroup>
-                </div>
-            </div>
-        );
-    }
-});
-
-var ProjectListRow = React.createClass({
-    render: function() {
-        return (
-            <div>
-                <div style={{width: 320, float: 'left'}}>
-                    {this.props.project.name}
-                </div>
-                <div style={{marginLeft: 20, width: 320, float: 'left'}}>
-                    {this.props.project.created}
-                </div>
-                <div style={{marginLeft: 710}}>
-                    <EditProjectForm
-                        clickEditProject={this.props.clickEditProject}
-                        project={this.props.project}
-                        projectDetails={this.props.projectDetails}
-                        handleInputChange={this.props.handleInputChange}
-                        handleSubmit={this.props.updateProjectInfo}
-                    />
-                        {/* Glyphicons don't work with npm bootstrap!
-                        <span className="glyphicon glyphicon-edit"
-                              title="Edit">
-                        </span>
-                        */}
-                    <a href="#" onClick={this.props.deleteProject.bind(null, this.props.project.id)}>
-                        {/* Glyphicons don't work with npm bootstrap!
-                        <span style={{marginLeft: 10}}
-                              className="glyphicon glyphicon-trash"
-                              title="Delete">
-                        </span>
-                        */}
-                        { " " }[Delete]
-                    </a>
-                </div>
-            </div>
-        );
-    }
-});
-
-const modalStyles = {
-    content : {
-        top                   : '50%',
-        left                  : '50%',
-        right                 : 'auto',
-        bottom                : 'auto',
-        marginRight           : '-50%',
-        transform             : 'translate(-50%, -50%)'
-    }
-};
-
-var EditProjectForm = React.createClass({
-    getInitialState: function() {
-        return {modalIsOpen: false};
-    },
-    clickEdit: function() {
-        this.props.clickEditProject(this.props.project.id);
-        this.openModal();
-    },
-    submit: function(e) {
-        this.props.handleSubmit(e);
-        this.closeModal();
-    },
-    openModal: function() {
-        this.setState({modalIsOpen: true});
-    },
-    afterOpenModal: function() {
-        // TODO
-    },
-    closeModal: function() {
-        this.setState({modalIsOpen: false});
-    },
-    render: function() {
-        return (
-            <span>
-                <a href="#"
-                   onClick={this.clickEdit}>
-                    [Edit]
-                </a>
-
-                <Modal
-                    isOpen={this.state.modalIsOpen}
-                    onAfterOpen={this.afterOpenModal}
-                    onRequestClose={this.closeModal}
-                    style={modalStyles} >
-
-                    <FormTitleRow formTitle="Edit Project" />
-                    <FormInputRow inputName="Project Name"
-                                  inputTag="input"
-                                  inputType="text"
-                                  formName="selectedProjectToEdit"
-                                  value={this.props.projectDetails["Project Name"]}
-                                  handleInputChange={this.props.handleInputChange}
-                    />
-                    <FormInputRow inputName="Description/notes"
-                                  inputTag="input"
-                                  inputType="text"
-                                  formName="selectedProjectToEdit"
-                                  value={this.props.projectDetails["Description/notes"]}
-                                  handleInputChange={this.props.handleInputChange}
-                    />
-                    <div className="submitButtonDiv" style={{marginTop: 15}}>
-                        <input type="submit"
-                               onClick={this.submit}
-                               value="Submit"
-                               className="submitButton"
-                        />
-                    </div>
-
-                </Modal>
-            </span>
-        );
-    }
-});
-
-var FormInputRow = React.createClass({
-    render: function() {
-        return (
-            <div className="formInputRow">
-                <div className="formInputTitle"
-                     style={{width: 320, float: 'left', marginTop: 5}}>
-                    {this.props.inputName}
-                </div>
-                <div className="formInputField"
-                     style={{marginLeft: 340, marginTop: 5}}>
-                    <this.props.inputTag
-                                 type={this.props.inputType}
-                                 value={this.props.value}
-                                 onChange={this.props.handleInputChange.bind(
-                                         null, this.props.inputName,
-                                         this.props.inputType,
-                                         this.props.formName)}
-                    />
-                </div>
-            </div>
-        );
-    }
-});
-
-var FormSelectInput = React.createClass({
-    render: function() {
-        var selectOptions = this.props.optionsList.map(function(option) {
-            return (
-                <option value={option.id} key={option.id}>
-                    {option.name}
-                </option>
-            );
-        }.bind(this));
-        return (
-            <div className="formInputRow">
-                <div className="formInputTitle"
-                     style={{width: 320, float: 'left', marginTop: 5}}>
-                    {this.props.inputName}
-                </div>
-                <div className="formInputField"
-                     style={{marginLeft: 340, marginTop: 5}}>
-                    <select
-                        value={this.props.value}
-                        onLoad={this.props.handleInputChange.bind(
-                                null, this.props.inputName,
-                                this.props.inputType,
-                                this.props.formName)}
-                        onChange={this.props.handleInputChange.bind(
-                                null, this.props.inputName,
-                                this.props.inputType,
-                                this.props.formName)}>
-                        {selectOptions}
-                    </select>
-                </div>
-            </div>
-        );
-    }
-});
-
-var FormTitleRow = React.createClass({
-    render: function() {
-        return (
-            <div className="formTitleDiv" style={{marginTop: 30}}>
-                <h3>
-                    {this.props.formTitle}
-                </h3>
-            </div>
-        );
-    }
-});
 
 var DatasetsTabContent = React.createClass({
     render: function() {
@@ -827,6 +610,8 @@ var FeatureSelectionDialog = React.createClass({
 });
 
 ReactDOM.render(
-    <MainContent />,
-    document.getElementById('content')
+  <Provider store={store}>
+    <MainContent />
+  </Provider>,
+  document.getElementById('content')
 );
